refactor(auth): refactor auth to assign get user encrypt algorith to it's right folder
