+++ packages/authentication/src/drivers/crypto/index.js
@@ -1,11 +1,9 @@
 const hmac = require('./hmacs');
 const hashMD5 = require('./hashes');
 const encryptString = require('./encrypt-string');
-const getUserEncryptAlgorithm = require('./get-user-encrypt-algorithm');
 
 module.exports = {
   hmac,
   hashMD5,
   encryptString,
-  getUserEncryptAlgorithm,
 };
