+++ packages/authentication/src/use-cases/crypto/index.js
@@ -1,0 +1,3 @@
+const getUserEncryptAlgorithm = require('./get-user-encrypt-algorithm');
+
+module.exports = { getUserEncryptAlgorithm };
